+++ IntraLib/include/Sound/InstrumentLibrary.h
@@ -36,7 +36,7 @@ struct MusicalInstruments
 	SynthesizedInstrument DrumSound2;
 	DrumInstrument Drums;
 
-	static SynthesizedInstrument CreateGuitar(int n=15, float c=128, float d=1.5, float e=0.75, float f=1, float freqMult=0.5f, float duration=0.7f, float volume=0.6f);
+	static SynthesizedInstrument CreateGuitar(size_t n=15, float c=128, float d=1.5, float e=0.75, float f=1, float freqMult=0.5f, float duration=0.7f, float volume=0.6f);
 
 };
